+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -29,6 +29,7 @@
 #include "OpForbiddenOverloadCheck.h"
 #include "OpLogicalPostfixCheck.h"
 #include "OpMixedIncrementDecrementCheck.h"
+#include "PureOverrideCheck.h"
 #include "StmtSwitchCaseParentCheck.h"
 #include "StmtSwitchDefaultBreakCheck.h"
 #include "StmtSwitchDefaultLastCheck.h"
@@ -77,6 +78,8 @@ public:
         "bsl-op-logical-postfix");
     CheckFactories.registerCheck<OpMixedIncrementDecrementCheck>(
         "bsl-op-mixed-increment-decrement");
+    CheckFactories.registerCheck<PureOverrideCheck>(
+        "bsl-pure-override");
     CheckFactories.registerCheck<StmtSwitchCaseParentCheck>(
         "bsl-stmt-switch-case-parent");
     CheckFactories.registerCheck<StmtSwitchDefaultBreakCheck>(